requirements.txt
line-bot-sdk==1.8.0
flask==1.1.1
gunicorn
pymongo
imgurpython==1.1.7
requests
numpy
matplotlib
pandas
datetime
beautifulsoup4==4.8.1
schedule
pandas_datareader
html5lib
yfinance==0.1.50
twder
selenium==3.8.0
